requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.39
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==3.1.3
